refactor(filters): change the signature of FeedFilterConfig::build (#7)
This PR changes the signature of the trait from:

``` rust
pub trait FeedFilterConfig: DeserializeOwned {
  type Filter: FeedFilter;

  async fn build(&self) -> Result<Self::Filter>;
}
```

to:

``` rust
pub trait FeedFilterConfig: DeserializeOwned {
  type Filter: FeedFilter;

  async fn build(self) -> Result<Self::Filter>;
}
```

The main reason to make the change is because the latter makes more
sense. The config is supposed to be consumed when built into a filter.
The former signature is confusing because it implies that the config may
be used for other purpose after the build, or that the config may be
built into a filter repeatedly. Either implication is not true about the
expected use of the type.

---------

Co-authored-by: Shou Ya <shouya@users.noreply.github.com>