feat(client): Caching requests to servers for feed (#12)
This PR implements a client that caches the get requests to the servers
when fetching feeds. It is used in places where `reqwest::Client` is
used. Currently there are two places:

- the client used to fetch the source (default ttl: 15 minutes)
- the full_text filter (default ttl: 12 hours)

This mechanism is expected to help in speeding up the fetching of feeds
especially for repeated full-text requests.

I'm also planning to employ this cache tests to insert custom fixtures
for certain web addresses for future feature tests.

---------

Co-authored-by: Shou Ya <shouya@users.noreply.github.com>