fix(config)!: change content_type from endpoint config (#17)
Previously the `content_type` in endpoint config is used to enforce the
interpretation of the request body in a certain content type. This
feature is useful on some feeds that reports wrong content types.

Later I discovered that this feature is more generally needed. For
example, in the upcoming merge filter, we may also want to assume the
content_type for the merged source. So in this PR, I moved the
`content_type` field from endpoint's config to the client's config. I
also renamed it to `assume_content_type` to make it more clear.

BREAKING CHANGE: This change will break the existing configuration if
you use `content_type` field in an endpoint. You need to update your
endpoint config as follows:

``` yaml
# old
endpoints:
  - path: /foo
    source: http://website/feed.xml
    content_type: application/atom+xml
    filters: ...

# new
endpoints:
  - path: /foo
    source: http://website/feed.xml
    client:
      assume_content_type: application/atom+xml
    filters: ...
```

---------

Co-authored-by: Shou Ya <shouya@users.noreply.github.com>