test: read from fixture directory (#19)
previous approach to read from fixture by inserting fixture files into
the cache of the client was not working for two reasons:

- the dummy client is only constructed for the source client, which
means other requests like the one used in `merge_feed` filter can't use
the fixtures.
- the cache has its own purpose and should be considered orthogonal to
the fixtures, as it's consider an internal logic of the client. It's not
a good idea to rely on the cache for providing fixtures for testing.

This commit changes the approach to read from fixture files by
requesting with `fixtures://` scheme, which is then handled by a custom
adapter that reads the file from the fixture directory. Note, the
`fixtures://` scheme is only handled in test environment.

---------

Co-authored-by: Shou Ya <shouya@users.noreply.github.com>