feat(inspector): webui to inspect the feeds (#22)
This PR implements a webui to inspect the endpoints defined in the
config file. Features:

- view the list of endpoints
- copy endpoint url
- view the list of filters and their config
- preview the processed feed content
- preview the feed with various numbers of filters skipped
- view the xml source code of the feed

The webui is embedded in the release binary, so it should be available
out of the box. It's accessible from the root url of the server, e.g.
`http://localhost:4080/`.

Note: At current stage, the webui is read-only. I don't yet have a
concrete plan on how to design an edit interface. I'm open to
suggestions.

---------

Co-authored-by: Shou Ya <shouya@users.noreply.github.com>