feat(inspector-ui): add a reload button (#30)
Add a reload button that reloads the config manually.

In the future I may implement this to be automatic, probably with a
websocket connection to the backend to detect reload events. For the
time being, I think this button is a simple and good enough solution.

In the passing I found a bug with the response in `_inspector/config` is
not being updated when the config is reloaded. I fixed that as well.

---------

Co-authored-by: Shou Ya <shouya@users.noreply.github.com>