refactor: FeedFilter trait (#46)
Prior to this change, the `FeedFilter` trait is defined as follows:

``` rust
pub trait FeedFilter {
  async fn run(&self, feed: &mut Feed) -> Result<()>;
}
```

Now I want to change the `FeedFilter` trait to be as follows:

``` rust
pub trait FeedFilter {
  async fn run(&self, ctx: &mut FilterContext, feed: Feed) -> Result<Feed>;
}
```

There are two major changes made to this interface.

First, the `run` method now takes a `FilterContext` that can be used to
store and retrieve state. The can `FilterContext` is used to influence
the behavior of filters. Currently no filter uses this context yet. But
I plan to put the base address from the request to the context so that
the `client` downstream can use this information to reconstruct absolute
url from relative urls. (c.f. #36)

Secondly, I modified the `feed` parameter to allow each filter to take
full ownership of the feed. The main reason is that I hope to define a
`FeedFilter` as to a **stateful endofunction on a Feed** (or more
precisely: an async, fallible, and stateful endofunction). I think the
new type more clearly indicates its nature. The original signature hints
that a filter somehow "modifies" a feed, which is slightly different
conceptually.

-------

In addition to the above changes, I also extracted the `FilterPipeline`
type into a standalone module. The concept was previously known as
`Filters`, but I dislike terrible naming.

---------

Co-authored-by: Shou Ya <shouya@users.noreply.github.com>