feat(filters): add `note` filter (#50)
This filter allows you to add comments. Just insert a `note: <your
comment>` line in the filter list.

The `note` filter does not affect the pipeline in any way, although it
does count as a filter in the `limit_filters` setting.

Example:

```yaml
- path: /foo.xml
  source: https://example.com/foo.xml
  note: This is the note for the endpoint
  filters:
    - full_text: {}
    - note: You can add a note filter to add comments to the pipeline.
    - merge: https://example.com/bar.xml
    - note: |
        This is a multiline note.
        It can be useful to add more detailed comments.
```

---------

An alternative approach I considered was to add a `note` field to all
filters. However, this approach is limited because many filters are
configured not as a struct. Often the case the config is just a single
literal value, or a list of values. It would be inconsistent to add
`note` field to those filters. I made an attempt in
2a387019113b3e322ad5194c54673c9f733d4519, but reverted it due to the
above reason.

---------

Co-authored-by: shouya <shouya@users.noreply.github.com>