refactor(server): optimize reload logic (#60)
The purpose of this branch is to support partial reloading.

Now we will only:

- Only rebuild endpoints whose config changes
- Only rebuild filters whose config changes

If an endpoint's config was untouched, we will reuse the existing
endpoint. Same for any filters in the filter pipeline.

There are several benefits with this approach:

- faster, more lightweight reload
- less disruption to caches (e.g. http client cache)
- the server does not need to be restarted on config change
- open up the possibility of notifying and reporting config error to
inspector-ui

Aside from some logging changes, this branch should cause no functional
difference.

------------

It's not trivial to implement this feature. Doing so requires some
architectural change in the server implementation. Previously the routes
are statically added to the `axum::Router`. Now since the endpoints can
be dynamically added and removed without restarting axum server, I have
to implement a custom handler depend on a global instance of current
config and endpoint state.

In addition, I made a more clear separation between `Error` and
`ConfigError`. Functions that are called during building the config into
the service will only raise `ConfigError`. The trait definition for
`FeedFilterConfig` has also been updated.

---------

Co-authored-by: shouya <shouya@users.noreply.github.com>