fix(filters): mark `client` optional for merge filter (#77)
Previously, the `client` and `filters` fields were required (but has a
default value). One So if the user don't specify the `client` and
`filters` fields, they'll default to a struct with the default values.

One side effect is that the inspector-ui will try render the all-empty
field for `client` even if the user didn't specify it `client` their
config.

With this change, I wrapped the `client` and `filters` fields with
`Option` box, so now the `client` field will appear as `null` when
rendered in json.