feat: `case_sensitive` option for `highlight` and `sanitize` filter (#82)
This PR adds a `case_sensitive` option to the `highlight` and `sanitize`
filters. By default, it is set to `false`.

When `case_sensitive` is `true` for the `highlight` filter, it will only
highlight matches that have the same case as the search term.

Similarly, `case_sensitive` controls whether the `from` field of the
`replace` operation in the `sanitize` filter matches the case. Note that
currently, it's not possible to specify a case-sensitive `remove`
operation. As a workaround, you
can use a case-sensitive `replace` operation with `to: ""`.

Example:

```yaml
highlight:
  keywords:
    - foo
    - bar
  bg_color: '#ffff00'
  case_sensitive: true

sanitize:
  - remove: "foo"
  - remove_regex: '\\d+'
  - replace:
      from: "bar"
      to: "baz"
      case_sensitive: true
  - replace_regex:
      from: '\\w+'
      to: "qux"
      case_sensitive: true