fix: merge feeds with extra namespaces (#90)
Previously, the `merge` filter only merges the items from the other
feeds. It fails in case where the merged feeds have different set of
namespaces. If the source feed doesn't have the necessary namespaces
used by the items from merged feeds, then the output xml will invalid.
The issue was first reported in
https://github.com/shouya/rss-funnel/issues/86.

This PR fixes the issue by merging the namespaces of the merged feeds
into the source feed.