feat: improve post body rendering on Web UI (#106)
Previously, the "Rendered" view was parsed from the raw XML feed on the
front-end. The logic on determining which field to show as posts' bodies
is arbitrary. Since #100, I changed the logic to recognize more post
body types. However, such change was not reflected on the Web UI. This
PR updates the Web UI to show the expected body content.

Besides, I made some visual tweaks to the Rendered view to show more
information including post's publication date, feed's description, etc.

In addition, The status bar will now show the number of posts on
successful fetches.