feat(filters): detect magnet links from feed entries (#122)
This PR adds a new filter `magnet` that extracts magnet links from the
body of the feed entry and store them in places bittorrent clients
recognizes.

For feeds in RSS format, the magnet links will be stored in the
`<enclosure>`. For feeds in Atom format, the magnet links will be stored
in a new `<link>`.

The filter has two configurations:

- `info_hash` (boolean, default: false): whether or not to detect info
hash by length without `magnet:` prefix. It's used when the magnet link
is not directly given in the body but the info hash is somehow
extractable.
- `override_existing` (boolean, default: false): if the entry already
contains a magnet link in its enclosure, whether or not to override it
with the detected one.