feat: htmx-based web ui (#147)
This new web ui is written with htmx and maud for the goal of making the
ui self-contained in the rust project. Which means it requires no
additional build-step for front-end. htmx does most of the heavy lifting
so the new ui is smaller and hopefully easier to maintain than the old.

Currently, the new ui has mostly reached feature parity with the old
inspector ui. New features can be expected in the future.

The new ui is enabled by default. You can disable it the same way by
customizing `RSS_FUNNEL_INSPECTOR_UI` environment variable.

For compatibility reasons, the old ui is still available through the
`/_inspector` route. This route will be removed in a future release. Any
feedback on the new ui is welcome.