refactor: improve source representation (#148)
Previously dynamic source is represented as `Option<Source>`. Upon
reflection, the case for dynamic source is not semantically the case for
the lack of source despite a null source is how dynamic source is
represented in config. Therefore, it is better to represent dynamic
source as a variant of `Source`.